KOKKOS: fix int members holding floating point values
Six members and locals declared int hold values that are not integers,
so each truncates on assignment:

  - fix vibmode/kk: boltz, the Boltzmann constant, truncates to 0, which
    zeroes evib or divides by zero on diffuse surfaces
  - fft2d and fft3d: the 1/(nfast*nslow) normalization truncates to 0,
    so the normalization pass zeroes the transform
  - fix emit/face: prefactor, the emission modulation factor from an
    equal-style variable
  - fix emit/surf: npcurrent, an Np that can legitimately be fractional;
    truncation also lets a negative value slip past the guard
  - compute surf/kk and compute boundary/kk: mvv2e

The two mvv2e members are benign today because mvv2e is 1.0 in the
supported unit systems, but they have the same shape and would break the
same way, so fix them here rather than leave them to be rediscovered.

+++ src/KOKKOS/compute_boundary_kokkos.h
@@ -265,7 +265,7 @@ void boundary_tally_kk(double dtremain,
 
 
  private:
-  int mvv2e;
+  double mvv2e;
   DAT::t_int_1d d_which;
 
   DAT::tdual_float_2d_lr k_myarray; // local accumulator array
